Cloud Save implemented
Cloud save is implemented and confirmed to work in my personal game. It
is still not perfect, as you can currently only pass ByteBuffers. I also
updates the libGDX sample to show how to implement the cloud save
feature. The identifiers of a test app I created are included to clear
any confusion about what kind of identifiers are required.